[CR] extract LinkOnCloseOutputStream to top-level class
Java-Style!  This involves creating 3 new files, breaking a previous import
statement, and many additional code changes...
