+++ clients/hadoopfs/src/main/java/io/lakefs/Constants.java
@@ -1,0 +1,5 @@
+package io.lakefs;
+
+class Constants {
+    public static final String SEPARATOR = "/";
+}
