+++ clients/hadoopfs/src/main/java/io/lakefs/ObjectLocation.java
@@ -1,0 +1,40 @@
+package io.lakefs;
+
+import org.apache.hadoop.fs.Path;
+
+class ObjectLocation {
+    private String repository;
+    private String ref;
+    private String path;
+
+    public String getRepository() {
+	return repository;
+    }
+
+    public void setRepository(String repository) {
+	this.repository = repository;
+    }
+
+    public String getRef() {
+	return ref;
+    }
+
+    public void setRef(String ref) {
+	this.ref = ref;
+    }
+
+    public String getPath() {
+	return path;
+    }
+
+    public void setPath(String path) {
+	this.path = path;
+    }
+
+    static String trimLeadingSlash(String s) {
+	if (s.startsWith(Constants.SEPARATOR)) {
+	    return s.substring(1);
+	}
+	return s;
+    }
+}
